Remove unused xmtp-handler code for cleaner structure
